Started refactoring UI to fit dynamic generation
Mostly name changes, method optimizations, and an improved folder structure.

Notable additions are the MapUI class used to store all the relevant UI elements for ease of access, and the CloudSpawner that will be used to spawn the emission display. It will also link all the relevant UI elements to the cloud.